moved console scripts to their own files in "/ebmlite/tools/"
